+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/conversion-tasks/church-supplemental-agreement.cy.ts
@@ -1,0 +1,110 @@
+import { ProjectBuilder } from "cypress/api/projectBuilder";
+import projectApi from "cypress/api/projectApi";
+import { checkAccessibilityAcrossPages } from "cypress/support/reusableTests";
+import taskListPage from "cypress/pages/projects/tasks/taskListPage";
+import { ProjectType } from "cypress/api/taskApi";
+import projectRemover from "cypress/api/projectRemover";
+import { rdoLondonUser } from "cypress/constants/cypressConstants";
+import taskPage from "cypress/pages/projects/tasks/taskPage";
+import { Logger } from "cypress/common/logger";
+import TaskHelper from "cypress/api/taskHelper";
+import { urnPool } from "cypress/constants/testUrns";
+
+const project = ProjectBuilder.createConversionProjectRequest({
+    urn: { value: urnPool.conversionTasks.spen },
+});
+let projectId: string;
+let taskId: string;
+const otherUserProject = ProjectBuilder.createConversionFormAMatProjectRequest({
+    urn: { value: urnPool.conversionTasks.grylls },
+    userAdId: rdoLondonUser.adId,
+});
+let otherUserProjectId: string;
+
+describe("Conversion tasks - Church supplemental agreement", () => {
+    before(() => {
+        projectRemover.removeProjectIfItExists(project.urn.value);
+        projectRemover.removeProjectIfItExists(otherUserProject.urn.value);
+        projectApi.createConversionProject(project).then((createResponse) => {
+            projectId = createResponse.value;
+            projectApi.getProject(project.urn.value).then((response) => {
+                taskId = response.body.tasksDataId.value;
+            });
+        });
+        projectApi.createMatConversionProject(otherUserProject).then((createResponse) => {
+            otherUserProjectId = createResponse.value;
+        });
+    });
+
+    beforeEach(() => {
+        cy.login();
+        cy.acceptCookies();
+        cy.visit(`projects/${projectId}/tasks/church_supplemental_agreement`);
+    });
+
+    it("should expand and collapse guidance details", () => {
+        taskPage
+            .clickDropdown("Help checking for changes")
+            .hasDropdownContent("Changes that personalise the model documents to a school or trust");
+    });
+
+    it("should submit the form and persist selections", () => {
+        Logger.log("Select some checkboxes and save");
+        taskPage
+            .hasCheckboxLabel("Signed by school or trust")
+            .tick()
+            .hasCheckboxLabel("Signed by diocese")
+            .tick()
+            .saveAndReturn();
+        taskListPage
+            .hasTaskStatusInProgress("Church supplemental agreement")
+            .selectTask("Church supplemental agreement");
+
+        Logger.log("Unselect same checkboxes and save");
+        taskPage
+            .hasCheckboxLabel("Signed by school or trust")
+            .isTicked()
+            .untick()
+            .hasCheckboxLabel("Signed by diocese")
+            .isTicked()
+            .untick()
+            .saveAndReturn();
+        taskListPage
+            .hasTaskStatusNotStarted("Church supplemental agreement")
+            .selectTask("Church supplemental agreement");
+        taskPage
+            .hasCheckboxLabel("Signed by school or trust")
+            .isUnticked()
+            .hasCheckboxLabel("Signed by diocese")
+            .isUnticked();
+    });
+
+    it("should show task status based on the checkboxes that are checked", () => {
+        cy.visit(`projects/${projectId}/tasks`);
+
+        TaskHelper.updateChurchSupplementalAgreement(taskId, ProjectType.Conversion, "notStarted");
+        cy.reload();
+        taskListPage.hasTaskStatusNotStarted("Church supplemental agreement");
+
+        TaskHelper.updateChurchSupplementalAgreement(taskId, ProjectType.Conversion, "notApplicable");
+        cy.reload();
+        taskListPage.hasTaskStatusNotApplicable("Church supplemental agreement");
+
+        TaskHelper.updateChurchSupplementalAgreement(taskId, ProjectType.Conversion, "inProgress");
+        cy.reload();
+        taskListPage.hasTaskStatusInProgress("Church supplemental agreement");
+
+        TaskHelper.updateChurchSupplementalAgreement(taskId, ProjectType.Conversion, "completed");
+        cy.reload();
+        taskListPage.hasTaskStatusCompleted("Church supplemental agreement");
+    });
+
+    it("Should NOT see the 'save and return' button for another user's project", () => {
+        cy.visit(`projects/${otherUserProjectId}/tasks/church_supplemental_agreement`);
+        taskPage.noSaveAndReturnExists();
+    });
+
+    it("Check accessibility across pages", () => {
+        checkAccessibilityAcrossPages();
+    });
+});