+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/conversions/revoke-a-directive-academy-order.cy.ts
@@ -1,0 +1,102 @@
+import { ProjectBuilder } from "cypress/api/projectBuilder";
+import { urnPool } from "cypress/constants/testUrns";
+import projectRemover from "cypress/api/projectRemover";
+import projectApi from "cypress/api/projectApi";
+import taskListPage from "cypress/pages/projects/tasks/taskListPage";
+import daoRevocation from "cypress/pages/projects/daoRevocation";
+import projectDetailsPage from "cypress/pages/projects/projectDetails/projectDetailsPage";
+
+const directiveAcademyOrderProject = ProjectBuilder.createConversionProjectRequest({
+    urn: { value: urnPool.conversion.stChads },
+    hasAcademyOrderBeenIssued: true,
+});
+let directiveAcademyOrderId: string;
+
+const academyOrderProject = ProjectBuilder.createConversionFormAMatProjectRequest({
+    urn: { value: urnPool.conversion.jessons },
+    hasAcademyOrderBeenIssued: false,
+});
+let academyOrderId: string;
+
+describe("Complete conversion projects tests", () => {
+    before(() => {
+        projectRemover.removeProjectIfItExists(directiveAcademyOrderProject.urn.value);
+        projectRemover.removeProjectIfItExists(academyOrderProject.urn.value);
+        projectApi.createConversionProject(directiveAcademyOrderProject).then((response) => {
+            directiveAcademyOrderId = response.value;
+        });
+        projectApi.createMatConversionProject(academyOrderProject).then((response) => {
+            academyOrderId = response.value;
+        });
+    });
+
+    beforeEach(() => {
+        cy.login();
+        cy.acceptCookies();
+    });
+
+    it("should be able to revoke a directive academy order project assigned to me", () => {
+        cy.visit(`projects/${directiveAcademyOrderId}/tasks`);
+        taskListPage.clickButton("Record DAO revocation");
+        daoRevocation
+            .continue()
+            .hasCheckboxLabel("I confirm a minister has approved this decision")
+            .tick()
+            .hasCheckboxLabel("I confirm I have sent the letter confirming the revocation decision")
+            .tick()
+            .hasCheckboxLabel("I confirm I have saved a copy of the letter to the school’s SharePoint folder")
+            .tick()
+            .continue()
+            .selectReasonWithDetails(
+                "School rated good or outstanding",
+                "The school has improved its Ofsted rating to good.",
+            )
+            .selectReasonWithDetails(
+                "Safeguarding concerns addressed",
+                "The previous safeguarding issues have been resolved.",
+            )
+            .selectReasonWithDetails("School closed or closing", "The school has closed due to low enrollment.")
+            .selectReasonWithDetails("Change to government policy", "New policies have made the order unnecessary.")
+            .continue()
+            .withMinisterName("Minister McMinisterface")
+            .continue()
+            .withDateOfDecision("15", "06", "2024")
+            .continue();
+
+        projectDetailsPage
+            .inOrder()
+            .summaryShows("Decision")
+            .hasValue("DAO revoked")
+            .summaryShows("Reasons")
+            .containsValue("School rated good or outstanding")
+            .containsValue("The school has improved its Ofsted rating to good.")
+            .containsValue("Safeguarding concerns addressed")
+            .containsValue("The previous safeguarding issues have been resolved.")
+            .containsValue("School closed or closing")
+            .containsValue("The school has closed due to low enrollment.")
+            .containsValue("Change to government policy")
+            .containsValue("New policies have made the order unnecessary.")
+            .summaryShows("Decision maker's role")
+            .hasValue("Minister")
+            .summaryShows("Minister's name")
+            .hasValue("Minister McMinisterface")
+            .summaryShows("Date of decision")
+            .hasValue("15 June 2024")
+            .clickButton("Record DAO revocation");
+
+        taskListPage.containsSuccessBannerWithMessage("DAO revocation recorded successfully");
+        cy.reload();
+        taskListPage
+            .containsImportantBannerWithMessage(
+                "",
+                "This project‘s Directive Academy Order was revoked on 15 June 2024.Only Service Support team members can make changes to this project.",
+            )
+            .doesntContain("Revoke a Directive Academy Order")
+            .buttonDoesNotExist("Record DAO revocation");
+    });
+
+    it("should not be able to revoke an academy order project", () => {
+        cy.visit(`projects/${academyOrderId}/tasks`);
+        taskListPage.doesntContain("Revoke a Directive Academy Order").buttonDoesNotExist("Record DAO revocation");
+    });
+});